♻️ Rename `join_github_app` to generic `app`
