+++ app/static/VERSION.txt
