+++ app/static/images/govuk-opengraph-image.png
