+++ app/templates/pages/invitation-sent.html
