+++ app/templates/pages/join-selection.html
