+++ app/templates/pages/multiple-invitations-sent.html
