+++ app/templates/pages/outside-collaborator.html
