+++ tests/views/test_views.py
@@ -4,14 +4,14 @@ from unittest.mock import MagicMock, patch
 
 from flask import get_flashed_messages
 
-import join_github_app
-from join_github_app.main.services.github_service import GithubService
+import app
+from app.main.services.github_service import GithubService
 
 
 class TestViews(unittest.TestCase):
     def setUp(self):
         self.github_service = MagicMock(GithubService)
-        self.app = join_github_app.create_app(self.github_service, False)
+        self.app = app.create_app(self.github_service, False)
         self.app.config["SECRET_KEY"] = "test_flask"
 
     def test_default(self):
@@ -75,7 +75,7 @@ class TestViews(unittest.TestCase):
             self.assertIn("Azure", str(response.data))
 
     @patch(
-        "join_github_app.main.routes.join.app_config",
+        "app.main.routes.join.app_config",
         new=SimpleNamespace(
             github=SimpleNamespace(
                 organisations=[
@@ -129,7 +129,7 @@ class TestCompletedRateLimit(unittest.TestCase):
 
         self.org = "some-org"
         self.github_service = MagicMock(GithubService)
-        self.app = join_github_app.create_app(self.github_service, True)
+        self.app = app.create_app(self.github_service, True)
 
     def test_rate_limit(self):
         # Send requests until you receive a 429 response
